scripts/requirements.txt
semver==2.13.0
ruamel.yaml==0.17.21
